Changed contents of descriptors and dockerfile to something valid.  Renamed write-api-service's swagger to something else
