+++ write-api-client/pom.xml
@@ -144,6 +144,11 @@
             <artifactId>swagger-java-client</artifactId>
             <version>1.2.0-alpha.6-SNAPSHOT</version>
         </dependency>
+        <dependency>
+            <groupId>io.dockstore</groupId>
+            <artifactId>swagger-java-client</artifactId>
+            <version>1.2.0-alpha.6-SNAPSHOT</version>
+        </dependency>
     </dependencies>
