+++ write-api-client/src/main/java/io.dockstore.client/cli/Add.java
@@ -9,12 +9,12 @@ import java.nio.file.Paths;
 import com.google.common.collect.Lists;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
-import io.swagger.client.ApiException;
-import io.swagger.client.api.GAGHoptionalwriteApi;
-import io.swagger.client.model.Tool;
-import io.swagger.client.model.ToolDescriptor;
-import io.swagger.client.model.ToolDockerfile;
-import io.swagger.client.model.ToolVersion;
+import io.swagger.client.write.ApiException;
+import io.swagger.client.write.api.GAGHoptionalwriteApi;
+import io.swagger.client.write.model.Tool;
+import io.swagger.client.write.model.ToolDescriptor;
+import io.swagger.client.write.model.ToolDockerfile;
+import io.swagger.client.write.model.ToolVersion;
 import json.Output;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
